OSGi project updated
Progress made in Test moved to OSGi project.
OSGI only supports double variables.
OSGI has a bug in text conversion

+++ Calculator/src/calculator/CalculatorInterface.java
@@ -9,11 +9,11 @@ public interface CalculatorInterface {
 	}
 
 
-	public int add(int num1, int num2);
+	public double add(double num1, double num2);
 
-    public int subtract(int num1, int num2);
+    public double subtract(double num1, double num2);
 
-    public int multiply(int num1, int num2);
+    public double multiply(double num1, double num2);
 
-    public int divide(int num1, int num2);
+    public double divide(double num1, double num2);
 }
