+++ Calculator/src/calculator/CalculatorService.java
@@ -4,22 +4,22 @@ package calculator;
 public class CalculatorService implements CalculatorInterface {
     
 	@Override
-    public int add(int num1, int num2) {
+    public double add(double num1, double num2) {
         return num1 + num2;
     }
 
 	@Override
-    public int subtract(int num1, int num2) {
+    public double subtract(double num1, double num2) {
         return num1 - num2;
     }
 
 	@Override
-    public int multiply(int num1, int num2) {
+    public double multiply(double num1, double num2) {
         return num1 * num2;
     }
 
 	@Override
-    public int divide(int num1, int num2) {
+    public double divide(double num1, double num2) {
         return num1 / num2;
     }
